Moved pipeline APIs to ingest namespace
