+++ client/rest-high-level/src/main/java/org/elasticsearch/client/RestHighLevelClient.java
@@ -191,6 +191,7 @@ public class RestHighLevelClient implements Closeable {
 
     private final IndicesClient indicesClient = new IndicesClient(this);
     private final ClusterClient clusterClient = new ClusterClient(this);
+    private final IngestClient ingestClient = new IngestClient(this);
     private final SnapshotClient snapshotClient = new SnapshotClient(this);
     private final TasksClient tasksClient = new TasksClient(this);
 
@@ -254,6 +255,15 @@ public class RestHighLevelClient implements Closeable {
      */
     public final ClusterClient cluster() {
         return clusterClient;
+    }
+
+    /**
+     * Provides a {@link IngestClient} which can be used to access the Ingest API.
+     *
+     * See <a href="https://www.elastic.co/guide/en/elasticsearch/reference/current/ingest.html">Ingest API on elastic.co</a>
+     */
+    public final IngestClient ingest() {
+        return ingestClient;
     }
 
     /**
