+++ client/rest-high-level/src/test/java/org/elasticsearch/client/ClusterClientIT.java
@@ -22,20 +22,12 @@ package org.elasticsearch.client;
 import org.elasticsearch.ElasticsearchException;
 import org.elasticsearch.action.admin.cluster.settings.ClusterUpdateSettingsRequest;
 import org.elasticsearch.action.admin.cluster.settings.ClusterUpdateSettingsResponse;
-import org.elasticsearch.action.ingest.GetPipelineRequest;
-import org.elasticsearch.action.ingest.GetPipelineResponse;
-import org.elasticsearch.action.ingest.PutPipelineRequest;
-import org.elasticsearch.action.ingest.DeletePipelineRequest;
-import org.elasticsearch.action.ingest.WritePipelineResponse;
 import org.elasticsearch.cluster.routing.allocation.decider.EnableAllocationDecider;
-import org.elasticsearch.common.bytes.BytesReference;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.common.unit.ByteSizeUnit;
-import org.elasticsearch.common.xcontent.XContentBuilder;
 import org.elasticsearch.common.xcontent.XContentType;
 import org.elasticsearch.common.xcontent.support.XContentMapValues;
 import org.elasticsearch.indices.recovery.RecoverySettings;
-import org.elasticsearch.ingest.PipelineConfiguration;
 import org.elasticsearch.rest.RestStatus;
 
 import java.io.IOException;
@@ -112,54 +104,5 @@ public class ClusterClientIT extends ESRestHighLevelClientTestCase {
         assertThat(exception.status(), equalTo(RestStatus.BAD_REQUEST));
         assertThat(exception.getMessage(), equalTo(
                 "Elasticsearch exception [type=illegal_argument_exception, reason=transient setting [" + setting + "], not recognized]"));
-    }
-
-    public void testPutPipeline() throws IOException {
-        String id = "some_pipeline_id";
-        XContentBuilder pipelineBuilder = buildRandomXContentPipeline();
-        PutPipelineRequest request = new PutPipelineRequest(
-            id,
-            BytesReference.bytes(pipelineBuilder),
-            pipelineBuilder.contentType());
-
-        WritePipelineResponse putPipelineResponse =
-            execute(request, highLevelClient().cluster()::putPipeline, highLevelClient().cluster()::putPipelineAsync);
-        assertTrue(putPipelineResponse.isAcknowledged());
-    }
-
-    public void testGetPipeline() throws IOException {
-        String id = "some_pipeline_id";
-        XContentBuilder pipelineBuilder = buildRandomXContentPipeline();
-        {
-            PutPipelineRequest request = new PutPipelineRequest(
-                id,
-                BytesReference.bytes(pipelineBuilder),
-                pipelineBuilder.contentType()
-            );
-            createPipeline(request);
-        }
-
-        GetPipelineRequest request = new GetPipelineRequest(id);
-
-        GetPipelineResponse response =
-            execute(request, highLevelClient().cluster()::getPipeline, highLevelClient().cluster()::getPipelineAsync);
-        assertTrue(response.isFound());
-        assertEquals(response.pipelines().get(0).getId(), id);
-        PipelineConfiguration expectedConfig =
-            new PipelineConfiguration(id, BytesReference.bytes(pipelineBuilder), pipelineBuilder.contentType());
-        assertEquals(expectedConfig.getConfigAsMap(), response.pipelines().get(0).getConfigAsMap());
-    }
-
-    public void testDeletePipeline() throws IOException {
-        String id = "some_pipeline_id";
-        {
-            createPipeline(id);
-        }
-
-        DeletePipelineRequest request = new DeletePipelineRequest(id);
-
-        WritePipelineResponse response =
-            execute(request, highLevelClient().cluster()::deletePipeline, highLevelClient().cluster()::deletePipelineAsync);
-        assertTrue(response.isAcknowledged());
     }
 }
