+++ client/rest-high-level/src/test/java/org/elasticsearch/client/documentation/ClusterClientDocumentationIT.java
@@ -186,220 +186,4 @@ public class ClusterClientDocumentationIT extends ESRestHighLevelClientTestCase 
             assertTrue(latch.await(30L, TimeUnit.SECONDS));
         }
     }
-
-    public void testPutPipeline() throws IOException {
-        RestHighLevelClient client = highLevelClient();
-
-        {
-            // tag::put-pipeline-request
-            String source =
-                "{\"description\":\"my set of processors\"," +
-                    "\"processors\":[{\"set\":{\"field\":\"foo\",\"value\":\"bar\"}}]}";
-            PutPipelineRequest request = new PutPipelineRequest(
-                "my-pipeline-id", // <1>
-                new BytesArray(source.getBytes(StandardCharsets.UTF_8)), // <2>
-                XContentType.JSON // <3>
-            );
-            // end::put-pipeline-request
-
-            // tag::put-pipeline-request-timeout
-            request.timeout(TimeValue.timeValueMinutes(2)); // <1>
-            request.timeout("2m"); // <2>
-            // end::put-pipeline-request-timeout
-
-            // tag::put-pipeline-request-masterTimeout
-            request.masterNodeTimeout(TimeValue.timeValueMinutes(1)); // <1>
-            request.masterNodeTimeout("1m"); // <2>
-            // end::put-pipeline-request-masterTimeout
-
-            // tag::put-pipeline-execute
-            WritePipelineResponse response = client.cluster().putPipeline(request); // <1>
-            // end::put-pipeline-execute
-
-            // tag::put-pipeline-response
-            boolean acknowledged = response.isAcknowledged(); // <1>
-            // end::put-pipeline-response
-            assertTrue(acknowledged);
-        }
-    }
-
-    public void testPutPipelineAsync() throws Exception {
-        RestHighLevelClient client = highLevelClient();
-
-        {
-            String source =
-                "{\"description\":\"my set of processors\"," +
-                    "\"processors\":[{\"set\":{\"field\":\"foo\",\"value\":\"bar\"}}]}";
-            PutPipelineRequest request = new PutPipelineRequest(
-                "my-pipeline-id",
-                new BytesArray(source.getBytes(StandardCharsets.UTF_8)),
-                XContentType.JSON
-            );
-
-            // tag::put-pipeline-execute-listener
-            ActionListener<WritePipelineResponse> listener =
-                new ActionListener<WritePipelineResponse>() {
-                    @Override
-                    public void onResponse(WritePipelineResponse response) {
-                        // <1>
-                    }
-
-                    @Override
-                    public void onFailure(Exception e) {
-                        // <2>
-                    }
-                };
-            // end::put-pipeline-execute-listener
-
-            // Replace the empty listener by a blocking listener in test
-            final CountDownLatch latch = new CountDownLatch(1);
-            listener = new LatchedActionListener<>(listener, latch);
-
-            // tag::put-pipeline-execute-async
-            client.cluster().putPipelineAsync(request, listener); // <1>
-            // end::put-pipeline-execute-async
-
-            assertTrue(latch.await(30L, TimeUnit.SECONDS));
-        }
-    }
-
-    public void testGetPipeline() throws IOException {
-        RestHighLevelClient client = highLevelClient();
-
-        {
-            createPipeline("my-pipeline-id");
-        }
-
-        {
-            // tag::get-pipeline-request
-            GetPipelineRequest request = new GetPipelineRequest("my-pipeline-id"); // <1>
-            // end::get-pipeline-request
-
-            // tag::get-pipeline-request-masterTimeout
-            request.masterNodeTimeout(TimeValue.timeValueMinutes(1)); // <1>
-            request.masterNodeTimeout("1m"); // <2>
-            // end::get-pipeline-request-masterTimeout
-
-            // tag::get-pipeline-execute
-            GetPipelineResponse response = client.cluster().getPipeline(request); // <1>
-            // end::get-pipeline-execute
-
-            // tag::get-pipeline-response
-            boolean successful = response.isFound(); // <1>
-            List<PipelineConfiguration> pipelines = response.pipelines(); // <2>
-            for(PipelineConfiguration pipeline: pipelines) {
-                Map<String, Object> config = pipeline.getConfigAsMap(); // <3>
-            }
-            // end::get-pipeline-response
-
-            assertTrue(successful);
-        }
-    }
-
-    public void testGetPipelineAsync() throws Exception {
-        RestHighLevelClient client = highLevelClient();
-
-        {
-            createPipeline("my-pipeline-id");
-        }
-
-        {
-            GetPipelineRequest request = new GetPipelineRequest("my-pipeline-id");
-
-            // tag::get-pipeline-execute-listener
-            ActionListener<GetPipelineResponse> listener =
-                new ActionListener<GetPipelineResponse>() {
-                    @Override
-                    public void onResponse(GetPipelineResponse response) {
-                        // <1>
-                    }
-
-                    @Override
-                    public void onFailure(Exception e) {
-                        // <2>
-                    }
-                };
-            // end::get-pipeline-execute-listener
-
-            // Replace the empty listener by a blocking listener in test
-            final CountDownLatch latch = new CountDownLatch(1);
-            listener = new LatchedActionListener<>(listener, latch);
-
-            // tag::get-pipeline-execute-async
-            client.cluster().getPipelineAsync(request, listener); // <1>
-            // end::get-pipeline-execute-async
-
-            assertTrue(latch.await(30L, TimeUnit.SECONDS));
-        }
-    }
-
-    public void testDeletePipeline() throws IOException {
-        RestHighLevelClient client = highLevelClient();
-
-        {
-            createPipeline("my-pipeline-id");
-        }
-
-        {
-            // tag::delete-pipeline-request
-            DeletePipelineRequest request = new DeletePipelineRequest("my-pipeline-id"); // <1>
-            // end::delete-pipeline-request
-
-            // tag::delete-pipeline-request-timeout
-            request.timeout(TimeValue.timeValueMinutes(2)); // <1>
-            request.timeout("2m"); // <2>
-            // end::delete-pipeline-request-timeout
-
-            // tag::delete-pipeline-request-masterTimeout
-            request.masterNodeTimeout(TimeValue.timeValueMinutes(1)); // <1>
-            request.masterNodeTimeout("1m"); // <2>
-            // end::delete-pipeline-request-masterTimeout
-
-            // tag::delete-pipeline-execute
-            WritePipelineResponse response = client.cluster().deletePipeline(request); // <1>
-            // end::delete-pipeline-execute
-
-            // tag::delete-pipeline-response
-            boolean acknowledged = response.isAcknowledged(); // <1>
-            // end::delete-pipeline-response
-            assertTrue(acknowledged);
-        }
-    }
-
-    public void testDeletePipelineAsync() throws Exception {
-        RestHighLevelClient client = highLevelClient();
-
-        {
-            createPipeline("my-pipeline-id");
-        }
-
-        {
-            DeletePipelineRequest request = new DeletePipelineRequest("my-pipeline-id");
-
-            // tag::delete-pipeline-execute-listener
-            ActionListener<WritePipelineResponse> listener =
-                new ActionListener<WritePipelineResponse>() {
-                    @Override
-                    public void onResponse(WritePipelineResponse response) {
-                        // <1>
-                    }
-
-                    @Override
-                    public void onFailure(Exception e) {
-                        // <2>
-                    }
-                };
-            // end::delete-pipeline-execute-listener
-
-            // Replace the empty listener by a blocking listener in test
-            final CountDownLatch latch = new CountDownLatch(1);
-            listener = new LatchedActionListener<>(listener, latch);
-
-            // tag::delete-pipeline-execute-async
-            client.cluster().deletePipelineAsync(request, listener); // <1>
-            // end::delete-pipeline-execute-async
-
-            assertTrue(latch.await(30L, TimeUnit.SECONDS));
-        }
-    }
 }
