+++ docs/java-rest/high-level/ingest/get_pipeline.asciidoc
@@ -1,14 +1,14 @@
-[[java-rest-high-cluster-get-pipeline]]
+[[java-rest-high-ingest-get-pipeline]]
 === Get Pipeline API
 
-[[java-rest-high-cluster-get-pipeline-request]]
+[[java-rest-high-ingest-get-pipeline-request]]
 ==== Get Pipeline Request
 
 A `GetPipelineRequest` requires one or more `pipelineIds` to fetch.
 
 ["source","java",subs="attributes,callouts,macros"]
 --------------------------------------------------
-include-tagged::{doc-tests}/ClusterClientDocumentationIT.java[get-pipeline-request]
+include-tagged::{doc-tests}/IngestClientDocumentationIT.java[get-pipeline-request]
 --------------------------------------------------
 <1> The pipeline id to fetch
 
@@ -17,21 +17,21 @@ The following arguments can optionally be provided:
 
 ["source","java",subs="attributes,callouts,macros"]
 --------------------------------------------------
-include-tagged::{doc-tests}/ClusterClientDocumentationIT.java[get-pipeline-request-masterTimeout]
+include-tagged::{doc-tests}/IngestClientDocumentationIT.java[get-pipeline-request-masterTimeout]
 --------------------------------------------------
 <1> Timeout to connect to the master node as a `TimeValue`
 <2> Timeout to connect to the master node as a `String`
 
-[[java-rest-high-cluster-get-pipeline-sync]]
+[[java-rest-high-ingest-get-pipeline-sync]]
 ==== Synchronous Execution
 
 ["source","java",subs="attributes,callouts,macros"]
 --------------------------------------------------
-include-tagged::{doc-tests}/ClusterClientDocumentationIT.java[get-pipeline-execute]
+include-tagged::{doc-tests}/IngestClientDocumentationIT.java[get-pipeline-execute]
 --------------------------------------------------
 <1> Execute the request and get back the response in a GetPipelineResponse object.
 
-[[java-rest-high-cluster-get-pipeline-async]]
+[[java-rest-high-ingest-get-pipeline-async]]
 ==== Asynchronous Execution
 
 The asynchronous execution of a get pipeline request requires both the `GetPipelineRequest`
@@ -40,7 +40,7 @@ method:
 
 ["source","java",subs="attributes,callouts,macros"]
 --------------------------------------------------
-include-tagged::{doc-tests}/ClusterClientDocumentationIT.java[get-pipeline-execute-async]
+include-tagged::{doc-tests}/IngestClientDocumentationIT.java[get-pipeline-execute-async]
 --------------------------------------------------
 <1> The `GetPipelineRequest` to execute and the `ActionListener` to use when
 the execution completes
@@ -54,13 +54,13 @@ A typical listener for `GetPipelineResponse` looks like:
 
 ["source","java",subs="attributes,callouts,macros"]
 --------------------------------------------------
-include-tagged::{doc-tests}/ClusterClientDocumentationIT.java[get-pipeline-execute-listener]
+include-tagged::{doc-tests}/IngestClientDocumentationIT.java[get-pipeline-execute-listener]
 --------------------------------------------------
 <1> Called when the execution is successfully completed. The response is
 provided as an argument
 <2> Called in case of failure. The raised exception is provided as an argument
 
-[[java-rest-high-cluster-get-pipeline-response]]
+[[java-rest-high-ingest-get-pipeline-response]]
 ==== Get Pipeline Response
 
 The returned `GetPipelineResponse` allows to retrieve information about the executed
@@ -68,7 +68,7 @@ The returned `GetPipelineResponse` allows to retrieve information about the exec
 
 ["source","java",subs="attributes,callouts,macros"]
 --------------------------------------------------
-include-tagged::{doc-tests}/ClusterClientDocumentationIT.java[get-pipeline-response]
+include-tagged::{doc-tests}/IngestClientDocumentationIT.java[get-pipeline-response]
 --------------------------------------------------
 <1> Check if a matching pipeline id was found or not.
 <2> Get the list of pipelines found as a list of `PipelineConfig` objects.
