+++ docs/java-rest/high-level/supported-apis.asciidoc
@@ -106,14 +106,19 @@ include::indices/put_template.asciidoc[]
 The Java High Level REST Client supports the following Cluster APIs:
 
 * <<java-rest-high-cluster-put-settings>>
-* <<java-rest-high-cluster-put-pipeline>>
-* <<java-rest-high-cluster-get-pipeline>>
-* <<java-rest-high-cluster-delete-pipeline>>
 
 include::cluster/put_settings.asciidoc[]
-include::cluster/put_pipeline.asciidoc[]
-include::cluster/get_pipeline.asciidoc[]
-include::cluster/delete_pipeline.asciidoc[]
+
+== Ingest APIs
+The Java High Level REST Client supports the following Ingest APIs:
+
+* <<java-rest-high-ingest-put-pipeline>>
+* <<java-rest-high-ingest-get-pipeline>>
+* <<java-rest-high-ingest-delete-pipeline>>
+
+include::ingest/put_pipeline.asciidoc[]
+include::ingest/get_pipeline.asciidoc[]
+include::ingest/delete_pipeline.asciidoc[]
 
 == Snapshot APIs
